--- a/Cargo.toml
+++ b/Cargo.toml
@@ -18,3 +18,4 @@ ndarray = { version = "0.15", git = "https://github.com/3Dpass/ndarray", branch 
 peroxide = { version = "0.30", git = "https://github.com/3Dpass/Peroxide", branch = "devel", default-features = false }
 sha2 = { version = "0.10.6", default-features = false }
 base16ct = { version = "0.2.0", default-features = false }
+heapless = { version = "0.8.0", default-features = false }