… Cargo.lock
… tests/objs/Car.obj
… tests/objs/Wolf.obj
… tests/objs/cat.obj
… tests/objs/characterlowpoly2.obj
… tests/objs/cow.obj
… tests/objs/deer.obj
… tests/objs/fish.obj
… tests/objs/fish2.obj
… tests/objs/garbageTruck.obj
… tests/objs/goat.obj
… tests/objs/hummerHX.obj
… tests/objs/low-poly-mill.obj
… tests/objs/low-poly-tree.obj
… tests/objs/rat.obj
… tests/objs/t_3dpass-576329.obj
… tests/objs/trafficLight.obj
… tests/objs/treeBirch.obj
… tests/objs/warehouse.obj
… tests/objs/waterTank.obj
… tests/objs/wolf2.obj