Simple CellSet optimization. Faster ~25% on my machine.


time ./pass3d_org -s 12 -g 8 -a grid2d_v3a -d 10 -i ./objs/cow.obj  >> /dev/null
real	0m8,194s
user	0m8,198s
sys	0m0,000s

time ./pass3d -s 12 -g 8 -a grid2d_v3a -d 10 -i ./objs/cow.obj  >> /dev/null
real	0m5,880s
user	0m5,883s
sys	0m0,000s